# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "allium-deck",
    "version": "0.1.0",
    "summary": "Deck recommendation for Project Sekai: exact DFS with branch and bound.",
    "description": "Build parameters are the engine's own contract; every key documented in docs/parameters.md is accepted here, in camelCase or snake_case. The service adds only the request envelope and its own ceilings on `limit` and `timeoutMs`.",
    "license": { "name": "MIT OR Apache-2.0" }
  },
  "servers": [{ "url": "/", "description": "This service" }],
  "paths": {
...
    "/v1/world-bloom/support-cards": {
      "post": {
        "summary": "World Bloom support deck bonuses",
        "description": "Per-card support bonus for a World Bloom chapter, strongest first.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SupportCardsRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cards with a support bonus.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "cardId": { "type": "integer" },
                      "bonus": { "type": "number" }
                    },
                    "required": ["cardId", "bonus"]
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Overloaded" }
        }
      }
    },
    "/v1/music/recommend": {
      "post": {
        "summary": "Rank songs for a fixed deck",
        "description": "Scores every song and difficulty against an already-chosen deck.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MusicRecommendRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Songs ranked by score." },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Overloaded" }
        }
      }
    },
...
    "/v1/area-items/recommend": {
      "post": {
        "summary": "Area item upgrade order",
        "description": "Ranks area item upgrades by power gained per coin for the given cards.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "region": { "$ref": "#/components/schemas/Region" },
                  "user": { "$ref": "#/components/schemas/User" },
                  "cardIds": { "type": "array", "items": { "type": "integer" } },
                  "multiUnitBonusEvaluation": { "$ref": "#/components/schemas/MultiUnitBonusEvaluation" }
                },
                "required": ["user", "cardIds"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Upgrades ranked by power per coin." },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Overloaded" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminToken": { "type": "http", "scheme": "bearer" }
    },
    "responses": {
      "Error": {
        "description": "The request was rejected.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Overloaded": {
        "description": "The search queue is full. Retry after the interval in the Retry-After header.",
        "headers": {
          "Retry-After": { "schema": { "type": "string" } }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "invalid_request",
                  "unknown_region",
                  "overloaded",
                  "queue_timeout",
                  "internal"
                ]
              },
              "message": { "type": "string" }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      },
      "Region": {
        "type": "string",
        "description": "A region loaded by this service. Defaults to the service's default region; see GET /v1/regions."
      },
      "User": {
        "description": "The player's card collection, either as an object or as its JSON text.",
        "oneOf": [{ "type": "object" }, { "type": "string" }]
      },
      "MultiUnitBonusEvaluation": {
        "type": "string",
        "enum": ["by_deck", "force_on", "force_off"],
        "default": "by_deck",
        "description": "Activation policy for owned multi-unit area effects. force_off retains the item's other effects. The snake_case request alias is also accepted."
      },
      "Params": {
        "type": "object",
        "description": "Build parameters. Every key in docs/parameters.md is accepted, in camelCase or snake_case; unknown keys are ignored. `limit` and `timeoutMs` are lowered to this service's ceilings, reported by GET /v1/regions.",
        "properties": {
          "multiUnitBonusEvaluation": { "$ref": "#/components/schemas/MultiUnitBonusEvaluation" },
          "fixedConstraintMode": {
            "type": "string",
            "enum": ["slots", "members"],
            "default": "slots",
            "description": "Ordered input slots by default. In finale recommendations, members requires participation without fixing the leader; forcedLeaderCharacterId remains independent. Mandatory members with exact targetBonusList tiers are not supported. The snake_case alias is fixed_constraint_mode."
          }
        },
        "additionalProperties": true
      },
...
      "SupportCardsRequest": {
        "type": "object",
        "properties": {
          "region": { "$ref": "#/components/schemas/Region" },
          "user": { "$ref": "#/components/schemas/User" },
          "params": { "$ref": "#/components/schemas/Params" },
          "supportMasterMax": { "type": "boolean", "default": false },
          "supportSkillMax": { "type": "boolean", "default": false },
          "filterOtherUnit": { "type": "boolean", "default": false }
        },
        "required": ["user"]
      },
      "MusicRecommendRequest": {
        "type": "object",
        "properties": {
          "region": { "$ref": "#/components/schemas/Region" },
          "deck": {
            "type": "object",
            "properties": {
              "totalPower": { "type": "integer" },
              "eventBonusRate": { "type": "number" },
              "supportDeckBonusRate": { "type": "number" },
              "cards": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "skillScoreUp": { "type": "number" },
                    "skillLifeRecovery": { "type": "number" }
                  }
                }
              }
            }
          },
          "liveType": { "$ref": "#/components/schemas/LiveType" },
          "eventType": {
            "type": "string",
            "enum": ["marathon", "cheerful_carnival", "world_bloom"]
          },
          "eventId": { "type": "integer" },
          "skillOrderChooseStrategy": {
            "type": "string",
            "enum": ["average", "max", "min", "specific"],
            "default": "average"
          },
          "specificSkillOrder": { "type": "array", "items": { "type": "integer" } },
          "multiLiveTeammateScoreUp": { "type": "integer" },
          "multiLiveTeammatePower": { "type": "integer" }
        },
        "required": ["deck", "liveType"]
      },
...
      "LiveType": {
        "type": "string",
        "enum": [
          "solo",
          "auto",
          "multi",
          "cheerful",
          "challenge",
          "challenge_auto",
          "mysekai"
        ]
      },
...
    }
  }
}